ENG-1318 refactoring
